// frontend/src/types/types.tsx
export type phoneType = {
  id: number;
  name: string;
  type: string;
  price: number;
  processor: string;
  memory: number;
  battery: number;
  OS: string;
  rating: number;
  warranty: number;
  description: string;
  remainingUnits: number;
  image: string;
};
export type cardType = {
  id: number;
  name: string;
  price: number;
  rating: number;
  remainingUnits: number;
  image: string;
};